remove unused modeactivator.py
