+++ wakepy/core/activationmanager.py
@@ -5,13 +5,23 @@ from queue import Queue
 from threading import Thread
 
 from .activationresult import ActivationResult
-from .modeactivator import activate
 
 if typing.TYPE_CHECKING:
     from typing import List, Optional, Type
 
     from .dbus import DbusAdapter, DbusAdapterTypeSeq
     from .method import Method
+
+
+def activate(
+    methods: List[Type[Method]],
+    queue_in: Queue,
+    queue_out: Queue,
+    prioritize: Optional[List[Type[Method]]] = None,
+    dbus_adapter: DbusAdapter | DbusAdapterTypeSeq | None = None,
+):
+    """dummy function
+    TODO: Replace with real implementation"""
 
 
 class ModeActivationManager:
